+++ changelog.md
@@ -3,6 +3,7 @@
 ## 7.3.0
 
 selectImage gets a third optional argument, multiple, if true, it will act like a `<select multiple>`
+in that case onChange will receive an array of values
 
 ## 7.2.0